* master: (42 commits)
  Bump constance to latest
  Add miner dynamic config for job preload images
  bump bittensor wallet; fix tests
  bump bittensor only as much as needed; remove bt-decode override (we're at 0.6 already and it's required)
  Add logging in case of failed /tmp cleanup
  fix ci: bump nox in faci
  fix ci: bump nox
  precommit - uv lock
  encouraging celery to recycle workers a bit faster - bittensor leaks a lot of memory
  use metagraph snapshot miners in receipts due to a memory leak
  store bt shield cert in wallet directory; disable cert upload locally
  receipts container needs the wallet now (bt shield)
  bump bt shield to 0.1.7 (and bittensor to ~9.1.0 to keep it in line with the shield)
  Update locks
  Update docs
  Import build in a method
  Add UT for client.py
  Fix volumes, add UT for job.py
  Move fallback to optional-dependencies
  Add support for input/outputs volumes
  ...

+++ executor/app/src/compute_horde_executor/executor/management/commands/run_executor.py
@@ -1,12 +1,10 @@
 import asyncio
 import base64
 import csv
-import io
 import json
 import logging
 import os
 import pathlib
-import random
 import re
 import shlex
 import shutil
@@ -14,12 +12,9 @@ import subprocess
 import tempfile
 import time
 import typing
-import zipfile
 
-import httpx
 import packaging.version
 import pydantic
-from asgiref.sync import sync_to_async
 from compute_horde.base.docker import DockerRunOptionsPreset
 from compute_horde.certificate import (
     check_endpoint,
@@ -48,20 +43,16 @@ from compute_horde.protocol_messages import (
 )
 from compute_horde.transport import AbstractTransport, WSTransport
 from compute_horde.utils import MachineSpecs
+from compute_horde_core.output_upload import OutputUploader, OutputUploadFailed
 from compute_horde_core.volume import (
     HuggingfaceVolume,
-    InlineVolume,
-    MultiVolume,
-    SingleFileVolume,
     Volume,
-    ZipUrlVolume,
+    VolumeDownloader,
+    VolumeDownloadFailed,
 )
 from django.conf import settings
 from django.core.management.base import BaseCommand
-from huggingface_hub import snapshot_download
 from pydantic import TypeAdapter
-
-from compute_horde_executor.executor.output_uploader import OutputUploader, OutputUploadFailed
 
 logger = logging.getLogger(__name__)
 
@@ -276,7 +267,7 @@ class JobResult(pydantic.BaseModel):
     specs: MachineSpecs | None = None
     error_type: V0JobFailedRequest.ErrorType | None = None
     error_detail: str | None = None
-    upload_results: dict[str, str] | None = None
+    upload_results: dict[str, str]
 
 
 def truncate(v: str) -> str:
@@ -390,94 +381,6 @@ class JobError(Exception):
         self.error_detail = error_detail
 
 
-class DownloadManager:
-    def __init__(self, concurrency=3, max_retries=3):
-        self.semaphore = asyncio.Semaphore(concurrency)
-        self.max_retries = max_retries
-
-    def download_from_huggingface(
-        self,
-        relative_path: pathlib.Path,
-        repo_id: str,
-        revision: str | None,
-        repo_type: str | None = None,
-        allow_patterns: str | list[str] | None = None,
-    ):
-        try:
-            snapshot_download(
-                repo_id=repo_id,
-                repo_type=repo_type,
-                revision=revision,
-                token=settings.HF_ACCESS_TOKEN,
-                local_dir=relative_path,
-                allow_patterns=allow_patterns,
-            )
-        except Exception as e:
-            logger.error(f"Failed to download model from Hugging Face: {e}")
-            raise JobError(
-                f"Failed to download model from Hugging Face: {e}",
-                V0JobFailedRequest.ErrorType.HUGGINGFACE_DOWNLOAD,
-                str(e),
-            ) from e
-
-    async def download(self, fp, url):
-        async with self.semaphore:
-            retries = 0
-            bytes_received = 0
-            backoff_factor = 1
-
-            while retries < self.max_retries:
-                headers = {}
-                if bytes_received > 0:
-                    headers["Range"] = f"bytes={bytes_received}-"
-
-                async with httpx.AsyncClient() as client:
-                    async with client.stream("GET", url, headers=headers) as response:
-                        if response.status_code == 416:  # Requested Range Not Satisfiable
-                            # Server doesn't support resume, start from the beginning
-                            fp.seek(0)
-                            bytes_received = 0
-                            continue
-                        elif response.status_code != 206 and bytes_received > 0:  # Partial Content
-                            # Server doesn't support resume, start from the beginning
-                            fp.seek(0)
-                            bytes_received = 0
-                            continue
-
-                        if (
-                            bytes_received == 0
-                            and (content_length := response.headers.get("Content-Length"))
-                            is not None
-                        ):
-                            # check size early if Content-Length is present
-                            if 0 < settings.VOLUME_MAX_SIZE_BYTES < int(content_length):
-                                raise JobError("Input volume too large")
-
-                        try:
-                            async for chunk in response.aiter_bytes():
-                                bytes_received += len(chunk)
-                                if 0 < settings.VOLUME_MAX_SIZE_BYTES < bytes_received:
-                                    raise JobError("Input volume too large")
-                                fp.write(chunk)
-                            return  # Download completed successfully
-                        except (httpx.HTTPError, OSError) as e:
-                            retries += 1
-                            if retries >= self.max_retries:
-                                raise e
-
-                            # Exponential backoff with jitter
-                            backoff_time = backoff_factor * (2 ** (retries - 1))
-                            jitter = random.uniform(
-                                0, 0.1
-                            )  # Add jitter to avoid synchronization issues
-                            backoff_time *= 1 + jitter
-                            await asyncio.sleep(backoff_time)
-
-                            backoff_factor *= 2  # Double the backoff factor for the next retry
-
-            raise JobError(f"Download failed after {self.max_retries} retries")
-
-
 def job_container_name(docker_objects_infix: str) -> str:
     return f"ch-{docker_objects_infix}-job"
 
@@ -499,7 +402,6 @@ class JobRunner:
         self.output_volume_mount_dir = self.temp_dir / "output"
         self.specs_volume_mount_dir = self.temp_dir / "specs"
         self.artifacts_mount_dir = self.temp_dir / "artifacts"
-        self.download_manager = DownloadManager()
 
         self.job_container_name = job_container_name(settings.EXECUTOR_TOKEN)
         self.nginx_container_name = nginx_container_name(settings.EXECUTOR_TOKEN)
@@ -589,6 +491,7 @@ class JobRunner:
                 stdout=ex.description,
                 stderr="",
                 artifacts={},
+                upload_results={},
                 error_type=ex.error_type,
                 error_detail=ex.error_detail,
             )
@@ -616,6 +519,7 @@ class JobRunner:
                 stdout="",
                 stderr="",
                 artifacts={},
+                upload_results={},
                 exit_status=None,
             )
 
@@ -689,6 +593,7 @@ class JobRunner:
                     stderr=msg,
                     exit_status=None,
                     artifacts={},
+                    upload_results={},
                 )
 
         return None
@@ -774,6 +679,7 @@ class JobRunner:
             if job_request.output_upload:
                 try:
                     output_uploader = OutputUploader.for_upload_output(job_request.output_upload)
+                    output_uploader.max_size_bytes = settings.OUTPUT_ZIP_UPLOAD_MAX_SIZE_BYTES
                     upload_results = await output_uploader.upload(self.output_volume_mount_dir)
                 except OutputUploadFailed as ex:
                     logger.warning(
@@ -820,11 +726,19 @@ class JobRunner:
             "sh",
             "-c",
             f"rm -rf {shlex.quote(root_for_remove.as_posix())}/*",
-            stdout=asyncio.subprocess.DEVNULL,
-            stderr=asyncio.subprocess.DEVNULL,
+            stdout=asyncio.subprocess.PIPE,
+            stderr=asyncio.subprocess.PIPE,
         )
-        await process.wait()
-        self.temp_dir.rmdir()
+        stdout, stderr = await process.communicate()
+        if process.returncode == 0:
+            self.temp_dir.rmdir()
+        else:
+            logger.error(
+                f"Failed to clean up {self.temp_dir.as_posix()}/: process exited with return code {process.returncode}\n"
+                "Stdout and stderr:\n"
+                f"{truncate(stdout.decode())}\n"
+                f"{truncate(stderr.decode())}\n"
+            )
         if self.is_streaming_job:
             process = await asyncio.create_subprocess_exec(
                 "docker", "network", "rm", self.job_network_name
@@ -840,74 +754,31 @@ class JobRunner:
                 shutil.rmtree(path)
 
         if volume is not None:
-            if isinstance(volume, InlineVolume):
-                await self._unpack_inline_volume(volume)
-            elif isinstance(volume, ZipUrlVolume):
-                await self._unpack_zip_url_volume(volume)
-            elif isinstance(volume, SingleFileVolume):
-                await self._unpack_single_file_volume(volume)
-            elif isinstance(volume, MultiVolume):
-                await self._unpack_multi_volume(volume)
-            elif isinstance(volume, HuggingfaceVolume):
-                await self._unpack_huggingface_volume(volume)
+            # TODO(mlech): Refactor this to not treat `HuggingfaceVolume` with a special care
+            volume_downloader = VolumeDownloader.for_volume(volume)
+            volume_downloader.max_size_bytes = settings.VOLUME_MAX_SIZE_BYTES
+            if volume_downloader.handles_volume_type() is HuggingfaceVolume:
+                if volume.token is None:
+                    volume.token = settings.HF_ACCESS_TOKEN
+                try:
+                    await volume_downloader.download(self.volume_mount_dir)
+                except VolumeDownloadFailed as exc:
+                    logger.error(f"Failed to download model from Hugging Face: {exc}")
+                    raise JobError(
+                        str(exc),
+                        V0JobFailedRequest.ErrorType.HUGGINGFACE_DOWNLOAD,
+                        exc.error_detail,
+                    ) from exc
             else:
-                raise NotImplementedError(f"Unsupported volume_type: {volume.volume_type}")
+                try:
+                    await volume_downloader.download(self.volume_mount_dir)
+                except VolumeDownloadFailed as exc:
+                    raise JobError(str(exc)) from exc
 
         chmod_proc = await asyncio.create_subprocess_exec(
             "chmod", "-R", "777", self.temp_dir.as_posix()
         )
         assert 0 == await chmod_proc.wait()
-
-    async def _unpack_inline_volume(self, volume: InlineVolume):
-        decoded_contents = base64.b64decode(volume.contents)
-        bytes_io = io.BytesIO(decoded_contents)
-        zip_file = zipfile.ZipFile(bytes_io)
-        extraction_path = self.volume_mount_dir
-        if volume.relative_path:
-            extraction_path /= volume.relative_path
-        zip_file.extractall(extraction_path.as_posix())
-
-    async def _unpack_zip_url_volume(self, volume: ZipUrlVolume):
-        with tempfile.NamedTemporaryFile() as download_file:
-            await self.download_manager.download(download_file, volume.contents)
-            download_file.seek(0)
-            zip_file = zipfile.ZipFile(download_file)
-            extraction_path = self.volume_mount_dir
-            if volume.relative_path:
-                extraction_path /= volume.relative_path
-            zip_file.extractall(extraction_path.as_posix())
-
-    async def _unpack_huggingface_volume(self, volume: HuggingfaceVolume):
-        with tempfile.NamedTemporaryFile():
-            extraction_path = self.volume_mount_dir
-            if volume.relative_path:
-                extraction_path /= volume.relative_path
-            await sync_to_async(self.download_manager.download_from_huggingface)(
-                relative_path=extraction_path,
-                repo_id=volume.repo_id,
-                revision=volume.revision,
-                repo_type=volume.repo_type,
-                allow_patterns=volume.allow_patterns,
-            )
-
-    async def _unpack_single_file_volume(self, volume: SingleFileVolume):
-        file_path = self.volume_mount_dir / volume.relative_path
-        file_path.parent.mkdir(parents=True, exist_ok=True)
-        with file_path.open("wb") as file:
-            await self.download_manager.download(file, volume.url)
-
-    async def _unpack_multi_volume(self, volume: MultiVolume):
-        for sub_volume in volume.volumes:
-            if isinstance(sub_volume, InlineVolume):
-                await self._unpack_inline_volume(sub_volume)
-            elif isinstance(sub_volume, ZipUrlVolume):
-                await self._unpack_zip_url_volume(sub_volume)
-            elif isinstance(sub_volume, SingleFileVolume):
-                await self._unpack_single_file_volume(sub_volume)
-            elif isinstance(sub_volume, HuggingfaceVolume):
-                await self._unpack_huggingface_volume(sub_volume)
-            else:
-                raise NotImplementedError(f"Unsupported sub-volume type: {type(sub_volume)}")
 
     async def get_job_volume(self) -> Volume | None:
         initial_volume = self.initial_job_request.volume
